Use mimetype to determine image state when viewing galleries
It would do this by using the is_valid_image() method of the file - this is a
relatively expensive process compared to just checking the mimetype of the file
from the record we already have loaded.

This change relies just on the mimetype, speeding up page loads considerably
for galleries with large numbers of images.