Track filename properly when flipping jpeg files
Without this, jpeg files that are flipped cause the user to see an error in the
image editing interface (the flip worked, the error is annoying and misleading
however). This prevents that from happening by properly tracking the change in
filename.

This also cleans up duplicated image setup that is in every image editing mode.